+++ remote/package.json
@@ -8,7 +8,7 @@
     "graceful-fs": "4.2.3",
     "http-proxy-agent": "^2.1.0",
     "https-proxy-agent": "^2.2.3",
-    "iconv-lite-umd": "0.6.2",
+    "iconv-lite-umd": "0.6.3",
     "jschardet": "2.1.1",
     "minimist": "^1.2.5",
     "native-watchdog": "1.3.0",
