+++ remote/web/package.json
@@ -3,6 +3,8 @@
 	"version": "0.0.0",
 	"dependencies": {
 		"semver-umd": "^5.5.7",
+		"iconv-lite-umd": "0.6.3",
+		"jschardet": "2.1.1",
 		"vscode-oniguruma": "1.3.1",
 		"vscode-textmate": "5.1.1",
 		"xterm": "4.7.0-beta.3",
