+++ src/vs/workbench/services/textfile/test/node/encoding/encoding.test.ts
@@ -5,7 +5,7 @@
 
 import * as assert from 'assert';
 import * as fs from 'fs';
-import * as encoding from 'vs/base/node/encoding';
+import * as encoding from 'vs/workbench/services/textfile/common/encoding';
 import * as terminalEncoding from 'vs/base/node/terminalEncoding';
 import * as streams from 'vs/base/common/stream';
 import * as iconv from 'iconv-lite-umd';
