+++ src/vs/workbench/services/textfile/test/node/encoding/fixtures/some.cp1252.txt
