+++ src/vs/workbench/services/textfile/test/node/encoding/fixtures/some.png.txt
