+++ src/vs/workbench/services/textfile/test/node/encoding/fixtures/some.xml.png
