+++ src/vs/workbench/services/textfile/test/node/encoding/fixtures/some_utf16le.css
