+++ src/vs/workbench/services/textfile/test/node/encoding/fixtures/some_utf8.css
