+++ src/vs/workbench/services/textfile/test/node/encoding/fixtures/utf16_be_nobom.txt
