+++ src/vs/workbench/services/textfile/test/node/encoding/fixtures/utf16_le_nobom.txt
